Use long int for cont_index

+++ atomic.c
@@ -32,7 +32,7 @@ extern inline double photoionization_crosssection(int element, int ion, int leve
 extern inline double get_phixs_threshold(int element, int ion, int level, int phixstargetindex);
 
 
-static int get_continuumindex_phixstargetindex(int element, int ion, int level, int phixstargetindex)
+static long get_continuumindex_phixstargetindex(int element, int ion, int level, int phixstargetindex)
 /// Returns the index of the continuum associated to the given level.
 {
   return elements[element].ions[ion].levels[level].cont_index - phixstargetindex;
@@ -51,7 +51,7 @@ static int get_phixtargetindex(const int element, const int ion, const int level
 }
 
 
-int get_continuumindex(int element, int ion, int level, int upperionlevel)
+long get_continuumindex(int element, int ion, int level, int upperionlevel)
 /// Returns the index of the continuum associated to the given level.
 {
   const int phixstargetindex = get_phixtargetindex(element, ion, level, upperionlevel);
